--- a/log/experimental_level/doc.go
+++ b/log/experimental_level/doc.go
@@ -2,11 +2,11 @@
 // definitely have breaking changes and may be deleted altogether. Be warned!
 //
 // To use the level package, create a logger as per normal in your func main,
-// and wrap it with level.New.
+// and wrap it with level.NewFilter.
 //
 //    var logger log.Logger
 //    logger = log.NewLogfmtLogger(os.Stderr)
-//    logger = level.New(logger, level.Allowed(level.AllowInfoAndAbove())) // <--
+//    logger = level.NewFilter(logger, level.AllowInfoAndAbove()) // <--
 //    logger = log.NewContext(logger).With("ts", log.DefaultTimestampUTC)
 //
 // Then, at the callsites, use one of the level.Debug, Info, Warn, or Error
@@ -19,9 +19,6 @@
 //    }
 //
 // The leveled logger allows precise control over what should happen if a log
-// event is emitted without a level key, or if a squelched level is used. Check
-// the Option functions for details. And, you can easily use non-default level
-// values: create new string constants for whatever you want to change, pass
-// them explicitly to the Allowed Option function, and write your own level.Foo-style
-// helper methods.
+// event is emitted without a level key, or if a squelched level is used.
+// Check the Option functions for details.
 package level