The main focus here is to get away from identifying level keyvals with string matching and use type matching with a bit set for level filtering in the manner demonstrated by @seh in #453.

I am very happy that all of the configuration options could be preserved. I have taken the liberty to simplify the `Options` API a bit to make it friendlier. The `Allowed` option was removed in favor of converting all of its helper functions (e.g. `AllowInfoAndAbove`) into functional Options that can be used directly. I also changed `level.New` to `level.NewFilter` because it better expresses its purpose.

The documentation is updated and corrected.

Before updating the implementation I rewrote the benchmarks to be more comprehensive with the use of the new (Go 1.7) sub-benchmark feature. The before and after results of the benchmarks follow.

Updated benchmarks after 6a12b3162160a71abf7458cdfccdc8aa0cca0ba0.
```
name                                              old time/op    new time/op    delta
/NoContext/Baseline/Nop-8                            434ns ±13%     371ns ± 7%  -14.55%  (p=0.032 n=5+5)
/NoContext/Baseline/Logfmt-8                         796ns ± 1%     828ns ± 1%   +3.99%  (p=0.008 n=5+5)
/NoContext/Baseline/JSON-8                          2.06µs ± 3%    2.06µs ± 2%     ~     (p=0.690 n=5+5)
/NoContext/DisallowedLevel/Nop-8                     422ns ± 6%     368ns ±19%     ~     (p=0.095 n=5+5)
/NoContext/DisallowedLevel/Logfmt-8                  435ns ± 9%     377ns ±12%     ~     (p=0.056 n=5+5)
/NoContext/DisallowedLevel/JSON-8                    422ns ±11%     355ns ± 4%  -15.89%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Nop-8                        447ns ± 8%     343ns ± 9%  -23.20%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Logfmt-8                     824ns ± 1%     833ns ± 2%     ~     (p=0.111 n=5+5)
/NoContext/AllowedLevel/JSON-8                      2.07µs ± 3%    2.05µs ± 1%     ~     (p=0.690 n=5+5)
/TimeContext/Baseline/Nop-8                          780ns ± 4%     702ns ± 1%   -9.95%  (p=0.008 n=5+5)
/TimeContext/Baseline/Logfmt-8                      1.46µs ± 2%    1.47µs ± 1%     ~     (p=0.222 n=5+5)
/TimeContext/Baseline/JSON-8                        3.00µs ± 1%    2.99µs ± 4%     ~     (p=0.690 n=5+5)
/TimeContext/DisallowedLevel/Nop-8                   799ns ± 3%     711ns ± 2%  -11.06%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Logfmt-8                796ns ± 3%     698ns ± 0%  -12.23%  (p=0.000 n=5+4)
/TimeContext/DisallowedLevel/JSON-8                  792ns ± 1%     714ns ± 3%   -9.85%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Nop-8                      806ns ± 5%     717ns ± 1%  -10.97%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Logfmt-8                  1.47µs ± 1%    1.48µs ± 2%     ~     (p=0.056 n=5+5)
/TimeContext/AllowedLevel/JSON-8                    2.99µs ± 3%    2.94µs ± 0%     ~     (p=0.206 n=5+5)
/CallerContext/Baseline/Nop-8                       1.03µs ± 3%    0.97µs ± 1%   -6.45%  (p=0.008 n=5+5)
/CallerContext/Baseline/Logfmt-8                    2.31µs ± 1%    2.33µs ± 1%     ~     (p=0.246 n=5+5)
/CallerContext/Baseline/JSON-8                      3.86µs ± 1%    3.88µs ± 1%     ~     (p=0.286 n=5+5)
/CallerContext/DisallowedLevel/Nop-8                1.04µs ± 0%    0.99µs ± 3%   -4.61%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/Logfmt-8             1.08µs ± 6%    0.97µs ± 2%  -10.10%  (p=0.016 n=5+4)
/CallerContext/DisallowedLevel/JSON-8               1.04µs ± 0%    1.14µs ±29%     ~     (p=1.000 n=4+5)
/CallerContext/AllowedLevel/Nop-8                   1.04µs ± 0%    1.06µs ± 4%     ~     (p=0.698 n=4+5)
/CallerContext/AllowedLevel/Logfmt-8                2.38µs ± 4%    2.79µs ±13%  +17.09%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/JSON-8                  4.03µs ± 3%    4.44µs ± 5%  +10.29%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Nop-8              1.57µs ±10%    1.57µs ± 8%     ~     (p=1.000 n=5+5)
/TimeCallerReqIDContext/Baseline/Logfmt-8           3.41µs ± 3%    3.56µs ± 1%   +4.56%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/JSON-8             5.60µs ± 2%    5.76µs ± 4%     ~     (p=0.056 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8       1.50µs ± 1%    1.46µs ± 2%   -2.69%  (p=0.032 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8    1.52µs ± 2%    1.46µs ± 1%   -3.86%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8      1.50µs ± 1%    1.46µs ± 2%   -2.67%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Nop-8          1.51µs ± 1%    1.46µs ± 1%   -3.11%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8       3.45µs ± 3%    3.49µs ± 1%     ~     (p=0.151 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/JSON-8         5.55µs ± 2%    5.80µs ± 2%   +4.42%  (p=0.008 n=5+5)

name                                              old alloc/op   new alloc/op   delta
/NoContext/Baseline/Nop-8                             288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/Baseline/Logfmt-8                          288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/Baseline/JSON-8                          1.06kB ± 0%    1.05kB ± 0%   -1.50%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/Nop-8                      288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/Logfmt-8                   288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/JSON-8                     288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Nop-8                         288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Logfmt-8                      288B ± 0%      256B ± 0%  -11.11%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/JSON-8                      1.06kB ± 0%    1.05kB ± 0%   -1.50%  (p=0.008 n=5+5)
/TimeContext/Baseline/Nop-8                           384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/Baseline/Logfmt-8                        384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/Baseline/JSON-8                        1.28kB ± 0%    1.26kB ± 0%   -1.25%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Nop-8                    384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Logfmt-8                 384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/JSON-8                   384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Nop-8                       384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Logfmt-8                    384B ± 0%      352B ± 0%   -8.33%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/JSON-8                    1.28kB ± 0%    1.26kB ± 0%   -1.25%  (p=0.008 n=5+5)
/CallerContext/Baseline/Nop-8                         352B ± 0%      320B ± 0%   -9.09%  (p=0.008 n=5+5)
/CallerContext/Baseline/Logfmt-8                      520B ± 0%      488B ± 0%   -6.15%  (p=0.008 n=5+5)
/CallerContext/Baseline/JSON-8                      1.42kB ± 0%    1.40kB ± 0%   -1.13%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/Nop-8                  352B ± 0%      320B ± 0%   -9.09%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/Logfmt-8               352B ± 0%      320B ± 0%   -9.09%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/JSON-8                 352B ± 0%      320B ± 0%   -9.09%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/Nop-8                     352B ± 0%      320B ± 0%   -9.09%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/Logfmt-8                  520B ± 0%      488B ± 0%   -6.15%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/JSON-8                  1.42kB ± 0%    1.40kB ± 0%   -1.13%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Nop-8                592B ± 0%      560B ± 0%   -5.41%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Logfmt-8             768B ± 0%      736B ± 0%   -4.17%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/JSON-8             1.86kB ± 0%    1.85kB ± 0%   -0.86%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8         592B ± 0%      560B ± 0%   -5.41%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8      592B ± 0%      560B ± 0%   -5.41%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8        592B ± 0%      560B ± 0%   -5.41%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Nop-8            592B ± 0%      560B ± 0%   -5.41%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8         768B ± 0%      736B ± 0%   -4.17%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/JSON-8         1.86kB ± 0%    1.85kB ± 0%   -0.86%  (p=0.008 n=5+5)

name                                              old allocs/op  new allocs/op  delta
/NoContext/Baseline/Nop-8                             9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/Baseline/Logfmt-8                          9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/Baseline/JSON-8                            24.0 ± 0%      23.0 ± 0%   -4.17%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/Nop-8                      9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/Logfmt-8                   9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/DisallowedLevel/JSON-8                     9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Nop-8                         9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/Logfmt-8                      9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/NoContext/AllowedLevel/JSON-8                        24.0 ± 0%      23.0 ± 0%   -4.17%  (p=0.008 n=5+5)
/TimeContext/Baseline/Nop-8                           10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/Baseline/Logfmt-8                        10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/Baseline/JSON-8                          28.0 ± 0%      27.0 ± 0%   -3.57%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Nop-8                    10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/Logfmt-8                 10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/DisallowedLevel/JSON-8                   10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Nop-8                       10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/Logfmt-8                    10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
/TimeContext/AllowedLevel/JSON-8                      28.0 ± 0%      27.0 ± 0%   -3.57%  (p=0.008 n=5+5)
/CallerContext/Baseline/Nop-8                         9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/CallerContext/Baseline/Logfmt-8                      13.0 ± 0%      11.0 ± 0%  -15.38%  (p=0.008 n=5+5)
/CallerContext/Baseline/JSON-8                        31.0 ± 0%      30.0 ± 0%   -3.23%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/Nop-8                  9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/Logfmt-8               9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/CallerContext/DisallowedLevel/JSON-8                 9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/Nop-8                     9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/Logfmt-8                  13.0 ± 0%      11.0 ± 0%  -15.38%  (p=0.008 n=5+5)
/CallerContext/AllowedLevel/JSON-8                    31.0 ± 0%      30.0 ± 0%   -3.23%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Nop-8                11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/Logfmt-8             16.0 ± 0%      14.0 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/Baseline/JSON-8               39.0 ± 0%      38.0 ± 0%   -2.56%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Nop-8         11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/Logfmt-8      11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/DisallowedLevel/JSON-8        11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Nop-8            11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/Logfmt-8         16.0 ± 0%      14.0 ± 0%  -12.50%  (p=0.008 n=5+5)
/TimeCallerReqIDContext/AllowedLevel/JSON-8           39.0 ± 0%      38.0 ± 0%   -2.56%  (p=0.008 n=5+5)
```