@peterbourgon 2017-02-24
While working on the proposal, it struck me that a great feature for NewFilter would be an option to promote unleveled log records to leveled, at a chosen level, on an opt-in basis. What do you think?