@ChrisHines 2017-02-25
>a great feature for NewFilter would be an option to promote unleveled log records to leveled

That might be useful to some people, but I think it should be implemented as a new `Logger` rather than an additional option on `NewFilter`. It is a different kind of functionality and orthogonal to what `NewFilter` does. `NewFilter` controls whether a log event is removed from the stream. This new idea is about modifying a log event by injecting a level keyval that doesn't exist.

#### Outline of one approach

My first idea was to export the level values and provide the following function.
```go
func NewInjector(logger log.Logger, level *levelValue) log.Logger
```
Which actually works even though `levelValue` is not exported, but it looks really weird. After futzing around a bit more I arrived at the following structure.

```go
func NewInjector(next log.Logger, level Value) log.Logger

type Value interface {
	String() string
	levelVal()
}

func Key() interface{}  { return key }
func ErrorValue() Value { return errorValue }
func WarnValue() Value  { return warnValue }
func InfoValue() Value  { return infoValue }
func DebugValue() Value { return debugValue }

var (
	key        interface{} = "level"
	errorValue             = &levelValue{level: levelError, name: "error"}
	warnValue              = &levelValue{level: levelWarn, name: "warn"}
	infoValue              = &levelValue{level: levelInfo, name: "info"}
	debugValue             = &levelValue{level: levelDebug, name: "debug"}
)

type levelValue struct {
	name string
	level
}

func (v *levelValue) String() string { return v.name }
func (v *levelValue) levelVal()      {}
```

Providing the `Key` and `[Level]Value` functions gives other packages access to the special values in a read only manner. This is valuable to users that want to write their own level filtering or injection logic. It gives an application that needs to modify the levels emitted by some library according to some specific rule the tools it needs to look for and assign level keyvals.

An application would use `NewInjector` as follows.
```go
var logger log.Logger
logger = log.NewLogfmtLogger(os.Stdout)
logger = level.NewInjector(logger, level.InfoValue())
```

I have this coded up with a unit test but no docs yet. Let me know if this looks worth adding and I'll finish it up and add it to this PR.