@peterbourgon 2017-02-25
Brilliant. Looks great to me. The `levelVal` method is just to distinguish the interface from e.g. `fmt.Stringer`, is that correct? And why are `func Key()` and `type key` of type `interface{}` versus e.g. `string`?