@seh 2017-02-25
I'm leery of the utility in exposing the level values like this, even behind these functions.

Though it sounds like you did consider an alternative like this, I'll describe it just to solicit your reaction. What if instead of exporting these six new symbols (a type and five functions), you instead exported four functions like this:
- `func WithDebugAsDefault(log.Logger) log.Logger`
- `func WithInfoAsDefault(log.Logger) log.Logger`
- `func WithWarnAsDefault(log.Logger) log.Logger`
- `func WithErrorAsDefault(log.Logger) log.Logger`

Each fixes `NewInjector`'s second parameter into the name of the function, and we avoid exposing the level values to callers.

We use the word "default" all the time in programming, but here it feels especially apt: for lack of a specified alternative, we use this value.