@peterbourgon 2017-02-25
Why are you leery of exposing level values? I think it's great: it means users may implement their own identical decorators, or even fall back to a more explicit behavior:

```go
logger.Log(level.Key, level.DebugValue, "msg", "hello")
```