@seh 2017-02-25
Doing so allows use of the level value in a key position, which one could emulate well enough with a bare string key, but it seems like the filtering `Logger` might have to contend with more weird situations pertaining to these values when the package doesn't have complete control over their insertion.

It's no worse than it was with strings as the keys and values, but with the private type we have the opportunity to constrain its use. It sounds like you see that more as an opportunity than a liability. I don't have any real examples of potential misuse here. That's why I said "leery," as opposed to "vehemently opposed." [The threat I described](https://github.com/go-kit/kit/pull/453#issuecomment-281223955) over in #453 is more likely to arise.