@ChrisHines 2017-02-25
I am not sympathetic to the concerns about misuse. This package provides convenient helpers that ensure the usage is correct. Code that chooses to use the individual values must also assume the responsibility of doing so correctly.

There is little we can do about code using string values that match the results of the `String` method on the levelValues. But it does raise an interesting question about what _defines_ a leveled log record, the typed data or the final `[]byte` representation.

I am also not sure how I feel about the fact that this PR will require packages to use its types in order to be compatible with its filter. That is not the case with the existing string based approach. Having said that, exporting the `Key` and `*Value` functions opens the door for people to write adapters for other logging packages (e.g. logrus, log15) or packages that use our `log.Logger` interface but some other means of adding level keyvals. The adapters can inspect the incoming data and convert it to this package's representation.