@ChrisHines 2017-02-25
>And why are func Key() and type key of type interface{} versus e.g. string?

That is where one of the allocations is saved. Preallocating those as interface values avoids an allocation per `Log` call. If they were strings then the compiler would add a heap allocation each time it gets converted to an interface{}. Passing a variable that is already an interface type, or some pointer type, to an interface{} parameter only copies the interface header, and doesn't need to copy the wrapped value and so doesn't need to allocate space for it.

Granted it does make the return type of `Key` look odd.