@seh 2017-02-25
> Granted it does make the return type of Key look odd.

You could convert it back and return a string from `Key`, though. That is, use the `interface{}` variables internally, but return a string to callers who want the key. Though that yields a less peculiar interface, it then forces each caller to pay a subtle price until they discover the optimization that we are using here internally.