@peterbourgon 2017-02-26
> Preallocating those as interface values avoids an allocation per Log call.

Perfect. Perhaps a comment to this effect?