@peterbourgon 2017-03-02
With that small addition to the comment on NewInjector, :shipit: